Refactor Character class and update CharacterDirection type
